Proper error propagation from tokenizer
Rather than panicking when encountering an invalid token, the tokenizer
now returns an error (still only a string, but it won't crash the
program anymore).

This commit illustrates the use of `transponse()` to deal with the fact
that the `Interator` trait's `next()` function always has to return an
`Option<T>`, it's really easy to use `transpose()` to get the error out
of the `Option` and deal with it first.