Define top level program AST node
This is to deal with the fact that 1) a program can be empty and 2) that
a program contains at most one expression. Previously this logic was in
the expression parsing code, but now that we can have expressions inside
other nodes (e.g. an expression between parentheses), the expression
parsing code can no longer make assumption about what comes after it.