Factor out code to parse left-associative operations
This CL extracts the logic to parse any left-associative binary operator
into a separate `parse_left_associative` function and adds a couple of
helper function for a bit better ergonomics.

I am not yet sure if the extra indirection is worth it, it significantly
increases the cognitive overhead to understand the code.